SignIn, SignUp, Forgot Password view added
