+++ edu_bridge_app/lib/controller.dart
@@ -1,3 +1,4 @@
+import 'package:edu_bridge_app/view_model/auth/forgot_password_controller.dart';
 import 'package:edu_bridge_app/view_model/auth/login_in_controller.dart';
 import 'package:edu_bridge_app/view_model/auth/sign_up_controller.dart';
 import 'package:get/get.dart';
@@ -5,7 +6,8 @@ import 'package:get/get.dart';
 class ControllerBinder extends Bindings {
   @override
   void dependencies() {
-    Get.put(LogInController());
+    Get.put(SignInController());
     Get.put(SignUpController());
+    Get.put(ForgotPasswordController());
   }
 }
