+++ edu_bridge_app/lib/resources/export.dart
@@ -4,10 +4,8 @@ export 'package:edu_bridge_app/resources/assets_path.dart';
 // Utils
 export 'package:edu_bridge_app/utils/widgets/custom_button.dart';
 export 'package:edu_bridge_app/utils/widgets/custom_text.dart';
-export 'package:edu_bridge_app/view/auth/login/login_view.dart';
 // Auth Views
 export 'package:edu_bridge_app/view/auth/pin_verification/pin_verification_view.dart';
-export 'package:edu_bridge_app/view/auth/sign_in/sign_in_view.dart';
 // View Models
 export 'package:edu_bridge_app/view_model/auth/login_in_controller.dart';
 //Controller
